requirements.txt
python-pptx
pandas
matplotlib